+++ Changelog.md
@@ -1,3 +1,12 @@
+### Development
+[Full Changelog](http://github.com/rspec/rspec-support/compare/v3.0.2...master)
+
+Bug Fixes:
+
+* Fix regression in `Support#method_handle_for` where proxy objects
+  with method delegated would wrongly not return a method handle.
+  (Jon Rowe, #90)
+
 ### 3.0.2 / 2014-06-20
 [Full Changelog](http://github.com/rspec/rspec-support/compare/v3.0.1...v3.0.2)